+++ common/src/main/java/ru/itis/tanks/game/model/Updatable.java
@@ -2,5 +2,5 @@ package ru.itis.tanks.game.model;
 
 public interface Updatable extends GameObject{
 
-    void update(long delta);
+    void update(int delta);
 }
